[client/rest]: refactor rosetta test utilities
 problem: data route tests can reuse construction route tests utilities
solution: extract common routines into rosettaTestUtils
